Transaction History Feature Added
Tri-Programmed:
@RoryEnright, @isharma-sharma, @slammyde7113

Purchase history for a specific customer can be queried now. Added an
ajax call to the transactions table in the api file, click handlers
in the events file, and success/failure responses in the ui. Transaction
history button added and a div was added to append the data to it.